third_party/blink/web_tests/fast/forms/interactive-validation-formnovalidate-child-expected.txt

Test if formnovalidate attribute works for a submit button with child elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS location.search.indexOf("i0=") != -1 is true
TEST COMPLETE
 Submit
